Remove features including SSO on wp-login/php
