+++ tests/testLockOptions.php
@@ -54,21 +54,6 @@ class TestLockOptions extends WP_Auth0_Test_Case {
 	}
 
 	/**
-	 * Test that the SSO options are built properly.
-	 */
-	public function testThatDefaultSsoOptionsAreCorrect() {
-		$lock_options = new WP_Auth0_Lock10_Options( [], self::$opts );
-
-		$sso_opts = $lock_options->get_sso_options();
-		$this->assertEquals( 'openid email profile', $sso_opts['scope'] );
-		$this->assertEquals( 'id_token', $sso_opts['responseType'] );
-		$this->assertEquals( 'http://example.org/index.php?auth0=implicit', $sso_opts['redirectUri'] );
-		$this->assertEquals( WP_Auth0_Nonce_Handler::get_instance()->get_unique(), $sso_opts['nonce'] );
-		$this->assertEquals( $lock_options->get_state_obj(), $sso_opts['state'] );
-		$this->assertArrayNotHasKey( 'authParams', $sso_opts );
-	}
-
-	/**
 	 * Test that the social_big_buttons option is not used.
 	 */
 	public function testThatSocialButtonStyleStaysBig() {
