+++ tests/testLoginManagerLogout.php
@@ -41,12 +41,11 @@ class TestLoginManagerLogout extends WP_Auth0_Test_Case {
 	}
 
 	/**
-	 * Test that logout does not redirect to Auth0 if SSO or SLO is not on.
+	 * Test that logout does not redirect to Auth0 if SLO is not on.
 	 */
-	public function testThatNothingHappensIfNotSsoSlo() {
+	public function testThatNothingHappensIfNotSlo() {
 		$this->startRedirectHalting();
 		self::auth0Ready( true );
-		self::$opts->set( 'sso', 0 );
 		self::$opts->set( 'singlelogout', 0 );
 		self::$opts->set( 'auto_login', 0 );
 
@@ -54,14 +53,13 @@ class TestLoginManagerLogout extends WP_Auth0_Test_Case {
 	}
 
 	/**
-	 * Test that a redirect to the Auth0 logout URL happens if SSO or SLO is turned on.
+	 * Test that a redirect to the Auth0 logout URL happens if SLO is turned on.
 	 */
-	public function testThatRedirectHappensIfSsoSlo() {
+	public function testThatRedirectHappensIfSlo() {
 		$this->startRedirectHalting();
 		self::$opts->set( 'domain', 'test.auth0.com' );
 		self::$opts->set( 'client_id', '__test_client_id__' );
 		self::$opts->set( 'client_secret', '__test_client_secret__' );
-		self::$opts->set( 'sso', 0 );
 		self::$opts->set( 'singlelogout', 1 );
 
 		$redirect_data_slo = [];
@@ -81,20 +79,6 @@ class TestLoginManagerLogout extends WP_Auth0_Test_Case {
 		$this->assertEquals( '/v2/logout', $logout_url_slo['path'] );
 		$this->assertContains( 'client_id=__test_client_id__', $logout_url_slo['query'] );
 		$this->assertContains( 'returnTo=' . rawurlencode( home_url() ), $logout_url_slo['query'] );
-
-		// Test that SSO has the same behavior.
-		self::$opts->set( 'sso', 1 );
-		self::$opts->set( 'singlelogout', 0 );
-
-		$redirect_data_sso = [];
-		try {
-			$this->login->logout();
-		} catch ( Exception $e ) {
-			$redirect_data_sso = unserialize( $e->getMessage() );
-		}
-
-		$this->assertEquals( 302, $redirect_data_sso['status'] );
-		$this->assertEquals( $redirect_data_slo['location'], $redirect_data_sso['location'] );
 	}
 
 	/**
@@ -105,7 +89,6 @@ class TestLoginManagerLogout extends WP_Auth0_Test_Case {
 		self::$opts->set( 'domain', 'test.auth0.com' );
 		self::$opts->set( 'client_id', '__test_client_id__' );
 		self::$opts->set( 'client_secret', '__test_client_secret__' );
-		self::$opts->set( 'sso', 0 );
 		self::$opts->set( 'singlelogout', 0 );
 		self::$opts->set( 'auto_login', 1 );
